+++ docs/changelog.rst
@@ -10,6 +10,9 @@ SlangPy uses a `semantic versioning <http://semver.org>`__ policy for its API.
 Version 0.31.0 (TBA)
 ----------------------------
 
+- Add support for vectorizing against Python lists
+- `NDBuffer` and `Tensor` `empty` / `zeros` apis consistent
+- Added `load_from_image` for `NDBuffer` and `Tensor`
 - Fix typings for ``float2x3``, ``float3x2``, ``float4x2`` and ``float4x3``.
 
 Version 0.30.0 (May 27, 2025)